+++ src/helpers/PullField.php
@@ -42,9 +42,9 @@ class PullField
     ];
 
     const PULL_ASSET_FIELDS = [
-        'seoImage' => ['fieldName' => 'seoImage', 'seoField' => 'seoImage', 'transformModeField' => 'seoImageTransformMode', 'transformName' => 'base'],
-        'ogImage' => ['fieldName' => 'ogImage', 'seoField' => 'seoImage', 'transformModeField' => 'ogImageTransformMode', 'transformName' => 'facebook'],
-        'twitterImage' => ['fieldName' => 'twitterImage', 'seoField' => 'seoImage', 'transformModeField' => 'twitterImageTransformMode', 'transformName' => 'twitter'],
+        ['fieldName' => 'seoImage', 'seoField' => 'seoImage', 'transformModeField' => 'seoImageTransformMode', 'transformName' => 'base'],
+        ['fieldName' => 'ogImage', 'seoField' => 'seoImage', 'transformModeField' => 'ogImageTransformMode', 'transformName' => 'facebook'],
+        ['fieldName' => 'twitterImage', 'seoField' => 'seoImage', 'transformModeField' => 'twitterImageTransformMode', 'transformName' => 'twitter'],
     ];
 
 
@@ -147,8 +147,6 @@ class PullField
             $source = $bundleSettings[$fieldName.'Source'] ?? '';
             $ids = $bundleSettings[$fieldName.'Ids'] ?? [];
             $sourceField = $bundleSettings[$fieldName.'Field'] ?? '';
-            $template = $bundleSettings[$fieldName.'Template'] ?? '';
-
             if (!empty($source)) {
                 $transformImage = $bundleSettings[$fieldName.'Transform'] ?? true;
                 $seoField = $fields['seoField'];
@@ -160,7 +158,6 @@ class PullField
                 // Quote all the things here for clarity
                 $transformName = '"'.$transformName.'"';
                 $transformMode = '"'.$transformMode.'"';
-                $template = '"'.$template.'"';
                 // Special-case Twitter transforms
                 if ($fieldName === 'twitterImage') {
                     $transformName = 'seomatic.helper.twitterTransform()';
@@ -308,14 +305,6 @@ class PullField
                                 $globalsSettings[$fieldNameHeight] = '{{ craft.app.assets.assetById('
                                     .$ids[0]
                                     .', '.$siteId.').height }}';
-                            }
-                            break;
-                        case 'fromTemplate':
-                            if (!empty($template)) {
-                                $globalsSettings[$fieldName] = '{{ seomatic.helper.socialImage('
-                                    .rtrim($objectPrefix.$elementName, '.')
-                                    .', ' . $transformName . ''
-                                    .', ' . $template . ') }}';
                             }
                             break;
                     }