+++ src/models/metatag/RobotsTag.php
@@ -42,6 +42,10 @@ class RobotsTag extends MetaTag
      */
     public function init()
     {
+        // Handle a bug where 'content' could be permanently set to 'none'
+        if ($this->content === 'none') {
+            $this->content = $this->environment['live']['content'] ?? '{seomatic.meta.robots}';
+        }
         parent::init();
     }