+++ src/services/MetaContainers.php
@@ -157,8 +157,8 @@ class MetaContainers extends Component
     {
         Craft::beginProfile('MetaContainers::loadMetaContainers', __METHOD__);
         // Avoid recursion
-        if (!Seomatic::$loadingContainers) {
-            Seomatic::$loadingContainers = true;
+        if (!Seomatic::$loadingMetaContainers) {
+            Seomatic::$loadingMetaContainers = true;
             $this->setMatchedElement($uri, $siteId);
             // Get the cache tag for the matched meta bundle
             $metaBundle = $this->getMatchedMetaBundle();
@@ -236,7 +236,7 @@ class MetaContainers extends Component
             }
             Seomatic::$seomaticVariable->init();
             MetaValueHelper::cache();
-            Seomatic::$loadingContainers = false;
+            Seomatic::$loadingMetaContainers = false;
         }
         Craft::endProfile('MetaContainers::loadMetaContainers', __METHOD__);
     }
@@ -551,6 +551,11 @@ class MetaContainers extends Component
                 $html .= $result;
             }
         }
+        // Special-case for requests for the MetaSiteVars "container"
+        if ($type === MetaSiteVars::CONTAINER_TYPE) {
+            $result = Json::encode($this->metaSiteVars->toArray());
+            $html .= $result;
+        }
 
         return $html;
     }
@@ -571,6 +576,11 @@ class MetaContainers extends Component
                 /** @noinspection SlowArrayOperationsInLoopInspection */
                 $htmlArray = array_merge($htmlArray, $metaContainer->renderArray());
             }
+        }
+        // Special-case for requests for the MetaSiteVars "container"
+        if ($type === MetaSiteVars::CONTAINER_TYPE) {
+            $result = Json::encode($this->metaSiteVars->toArray());
+            $htmlArray = array_merge($htmlArray, $this->metaSiteVars->toArray());
         }
 
         return $htmlArray;